pkg/container: move test registry implementation out of tests
Move the toy implementation of a container registry out of the
container_test package and into internal/testregistry so it can be
reused across subpackages of the repository.
